pyproject.toml
[project]
name = "deriv-agenticai-for-fraudops"
version = "0.1.0"
description = "Financial Intelligence Swarm for AI-powered fraud detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Agent orchestration
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    # Databases
    "neo4j>=5.0.0",
    "qdrant-client>=1.12.0",
    "mem0ai>=1.0.3",
    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    # ISO 20022 XML parsing (python-iso20022 removed due to version conflicts)
    "xsdata>=24.0",
    "xmltodict>=0.14.0",
    # PDF parsing
    "docling>=2.0.0",
    # Embeddings
    "fastembed>=0.4.0",
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "rich>=14.3.2",
    "httpx>=0.28.1",
    "langchain-aws>=1.2.2",
    "boto3>=1.42.42",
    # PDF generation for SAR reports
    "reportlab>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
